mcp-toolbox/requirements.txt
psycopg2-binary>=2.9.7
PyYAML>=6.0
python-dotenv>=1.0.0
